feature(telemetry): search  box fixes
- Load filters in the server component
- Load filters from search params
- Fix not_* filters
- Add trace id filter
- Show suggestions on focus
